# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Vea Backend API
  version: 2.1.0
  description: |
    Vea 是一个代理管理服务的 HTTP API，提供 FRouter、配置、Geo 资源、核心组件与代理运行配置（ProxyConfig）的统一管理。

    ## 特性
    - FRouter 管理：导入、测速、延迟测试、链路封装
    - 配置管理：导入 Xray JSON 或订阅链接，自动刷新
    - Geo 资源：GeoIP/GeoSite 管理
    - 核心组件：Xray/sing-box/Geo 核心下载与安装
    - 代理运行配置：ProxyConfig（入站模式、引擎选择、绑定 FRouter）
    - 系统设置：系统代理、前端设置
    - TUN 能力检查与配置：/tun/check、/tun/setup

    ## 版本兼容性
    - 本仓库不承诺向后兼容；API/状态 schema 可能发生破坏性变更
    - 如需稳定集成，请固定到具体版本/提交
  contact:
    name: Vea Project
    url: https://github.com/FlowerRealm/Vea
  license:
    name: MIT
    url: https://github.com/FlowerRealm/Vea/blob/main/LICENSE
servers:
  - url: http://localhost:19080
    description: 本地开发服务器
  - url: http://127.0.0.1:19080
    description: 本地环回地址
...
paths:
...
  /proxy/config:
    get:
      tags: [proxy]
      summary: 获取代理运行配置
      description: 获取当前代理运行配置（单例）
      operationId: getProxyConfig
      responses:
        '200':
          description: 成功返回代理运行配置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyConfig'

    put:
      tags: [proxy]
      summary: 更新代理运行配置
      description: 更新代理运行配置（单例）
      operationId: updateProxyConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyConfig'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProxyConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
  /proxy/start:
    post:
      tags: [proxy]
      summary: 启动代理（以 FRouter 为中心）
      description: 使用当前配置启动代理；请求体可覆盖部分字段（常见只传 frouterId）
      operationId: startProxy
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProxyConfig'
      responses:
        '200':
          description: 启动成功
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /tun/setup:
    post:
      tags: [tun]
      summary: 配置 TUN
      description: 执行一次性 TUN 配置（不同平台实现不同）
      operationId: setupTUN
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: 配置成功
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
...
  schemas:
...
    ProxyConfig:
      type: object
      description: 代理运行配置（单例）
      properties:
        inboundMode:
          type: string
          enum: [socks, http, mixed, tun]
        inboundPort:
          type: integer
        inboundConfig:
          type: object
          additionalProperties: true
        tunSettings:
          type: object
          additionalProperties: true
        resolvedService:
          type: object
          additionalProperties: true
        dnsConfig:
          type: object
          additionalProperties: true
        logConfig:
          type: object
          additionalProperties: true
        performanceConfig:
          type: object
          additionalProperties: true
        xrayConfig:
          type: object
          additionalProperties: true
        preferredEngine:
          type: string
          enum: [xray, singbox, auto]
        frouterId:
          type: string
        updatedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误信息
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...